refactor: split useKitEditorLogic into focused concern hooks
useKitEditorLogic was a 467-line god-hook mixing scan state machines,
a global keyboard handler, and error-forwarding effects in with its
actual job of composing the kit editor's sub-hooks. Extract the three
inline concerns into focused hooks in the same directory:

- useKitScanning: filesystem rescan + in-memory voice name inference,
  owning the scan-status and voice-flash feedback state and timers
  (shared via scheduleStatusClear/flashVoicePanels helpers instead of
  the duplicated timer blocks). ScanStatus moves here and is
  re-exported from useKitEditorLogic.
- useKitEditorKeyboardNav: the global keydown shortcuts (kit nav,
  scan, sequencer toggle, sample nav/preview), taking narrow callback
  params instead of whole sub-hook objects.
- useKitErrorReporting: the playback/kit/SampleWaveformError
  forwarding effects.

useKitEditorLogic keeps an identical return shape and drops to a
~250-line orchestrator that owns only glue state (selection,
sequencer open/focus) and the parent-callback wrappers. No callers
change; the existing useKitEditorLogic and KitEditor test suites (40
tests) pass unmodified, exercising the extracted hooks through the
real composition.

First slice of the audit's mega-hook backlog (next candidates:
useLedVisualization, useLocalStoreWizard).
